mcp-project/pyproject.toml
[project]
authors = [
	{name = "Shay Harding", email = "message@shayharding.me"}
]
classifiers = [
    "Private :: Do Not Upload",
	"Development Status :: 4 - Beta",
    "Environment :: Console",
	"Intended Audience :: Developers",
	"Intended Audience :: Information Technology",
	"License :: OSI Approved :: MIT License",
	"Programming Language :: Python :: 3",
	"Programming Language :: Python :: 3.12",
]
name = "mcp"
version = "0.1.0"
description = "MCP project for scraping websites"
keywords = ["mcp", "scraping", "websites"]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.12.15",
    "anthropic>=0.64.0",
    "bs4>=0.0.2",
    "firecrawl-py>=2.16.5",
    "gradio[mcp]>=4.44.1",
    "ipykernel>=6.30.1",
    "mcp[cli]>=1.10.1",
    "textblob>=0.19.0",
    "python-dotenv>=1.2.1",
    "firecrawl-py>=4.12.0",
]
[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.17.1",
    "pytest>=8.4.2",
    "pytest-cov>=6.3.0",
    "ruff>=0.12.12",
]

[tool.pytest.ini_options]
minversion = "8.4"
addopts = [
	"--cov=server",
	"--cov-report=html",
	"--cov-fail-under=85",
	"--strict-markers",
	"--strict-config",
	"-ra",
]
testpaths = ["tests"]
pythonpath = ["."]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["."]
branch = true
omit = [
    "tests/*"
]

[tool.coverage.report]
exclude_also = [
    "def __repr__",
    "if self\\.debug",
    "if settings\\.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "@(abc\\.)?abstractmethod",
]

## Linter
[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    "E",  	# pycodestyle errors
    "W",  	# pycodestyle warnings
    "F",  	# pyflakes
    "I",  	# isort
    "B",  	# flake8-bugbear
    "C4", 	# flake8-comprehensions
    "UP", 	# pyupgrade
    "ARG", 	# flake8-unused-arguments
    "SIM", 	# flake8-simplify
    "TC", 	# flake8-type-checking
    "TID", 	# flake8-tidy-imports
    "Q",   	# flake8-quotes
    "PL",  	# pylint
    "PT",  	# flake8-pytest-style
    "RUF", 	# ruff-specific rules
]
ignore = [
    "E501",     # line too long, handled by black
    "PLR0913",  # too many arguments
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs}/*" = ["E402","PLR0133"]

## Formatters
[tool.black]
target-version = ['py312']
line-length = 120
required-version = 25

[tool.isort]
profile = "black"
line_length = 120
skip = [".gitignore", ".dockerignore"]

## Typing
[tool.mypy]
python_version = "3.12"
exclude_gitignore = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true
warn_unreachable = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
check_untyped_defs = true
strict_equality = true
